created services and reformed rest controllers

+++ src/main/java/com/cinema/repo/HallRepository.java
@@ -3,5 +3,5 @@ package com.cinema.repo;
 import com.cinema.models.Hall;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HallRepository extends JpaRepository<Hall, Integer> {
+public interface HallRepository extends JpaRepository<Hall, Long> {
 }
